Add ProductTypeSurCharges

+++ src/Insurance.Api/Models/InsuranceDto.cs
@@ -5,6 +5,6 @@ namespace Insurance.Api.Models
     public class InsuranceDto
     {        
         public int ProductId { get; set; }
-        public float InsuranceValue { get; set; }
+        public decimal InsuranceValue { get; set; }
     }
 }
